FE-Fix: teams페이지 커스텀훅 수정 및 파일 하나로 구성
