+++ apps/web/app/admin/teams/_components/TeamListItem.tsx
@@ -9,8 +9,7 @@ import { type TeamType } from "@repo/types";
 import Dropdown from "@ui/src/components/common/Dropdown";
 import { Chevron } from "@ui/public";
 import useIsMobileStore from "@/app/store/useIsMobileStore";
-import { useDeleteTeam } from "../_hooks/useDeleteTeam";
-import { useUpdateTeam } from "../_hooks/useUpdateTeam";
+import { useDeleteTeam, useUpdateTeam } from "../_hooks/useTeamsMutations";
 import ManageTeamModal from "./ManageTeamModal";
 
 interface TeamListItemProps {
