--- apps/web/app/admin/teams/_hooks/useCreateForm.ts
@@ -1,10 +1,0 @@
-import { type FieldValues, useForm } from "react-hook-form";
-
-export const useCreateForm = (): FieldValues => {
-  return useForm({
-    mode: "onChange",
-    defaultValues: {
-      teamName: "",
-    },
-  });
-};
